When and WhenNot
When an pattern was founded by one string or regexp
one function was executed.

When not an pattern was founded by one string or regexp
one function was executed.

Method end has the job to execute one function and not
more than that.

Git ignore was added, to exclude the dir node_modules and
vim buffer files.